+++ LibreNMS/Authentication/ADAuthorizationAuthorizer.php
@@ -8,6 +8,7 @@ use LibreNMS\Exceptions\AuthenticationException;
 class ADAuthorizationAuthorizer extends MysqlAuthorizer
 {
     use LdapSessionCache;
+    use ADUtils;
 
     protected static $AUTH_IS_EXTERNAL = 1;
     protected static $CAN_UPDATE_PASSWORDS = 0;
@@ -73,7 +74,7 @@ class ADAuthorizationAuthorizer extends MysqlAuthorizer
         $search = ldap_search(
             $this->ldap_connection,
             Config::get('auth_ad_base_dn'),
-            ActiveDirectoryAuthorizer::userFilter($username),
+            $this->userFilter($username),
             array('samaccountname')
         );
         $entries = ldap_get_entries($this->ldap_connection, $search);
@@ -104,7 +105,7 @@ class ADAuthorizationAuthorizer extends MysqlAuthorizer
         $search = ldap_search(
             $this->ldap_connection,
             Config::get('auth_ad_base_dn'),
-            ActiveDirectoryAuthorizer::userFilter($username),
+            $this->userFilter($username),
             array('memberOf')
         );
         $entries = ldap_get_entries($this->ldap_connection, $search);
@@ -136,137 +137,21 @@ class ADAuthorizationAuthorizer extends MysqlAuthorizer
         $search = ldap_search(
             $this->ldap_connection,
             Config::get('auth_ad_base_dn'),
-            ActiveDirectoryAuthorizer::userFilter($username),
+            $this->userFilter($username),
             $attributes
         );
         $entries = ldap_get_entries($this->ldap_connection, $search);
 
         if ($entries['count']) {
-            $user_id = preg_replace('/.*-(\d+)$/', '$1', $this->sidFromLdap($entries[0]['objectsid'][0]));
+            $user_id = $this->getUseridFromSid($this->sidFromLdap($entries[0]['objectsid'][0]));
         }
 
         $this->authLdapSessionCacheSet('userid', $user_id);
         return $user_id;
     }
 
-    public function getUserlist()
+    protected function getConnection()
     {
-        $userlist = array();
-        $userhash = array();
-
-        $ldap_groups = $this->getGroupList();
-
-        foreach ($ldap_groups as $ldap_group) {
-            $search_filter = "(&(memberOf:1.2.840.113556.1.4.1941:=$ldap_group)(!(userAccountControl:1.2.840.113556.1.4.803:=2)))";
-            if (Config::get('auth_ad_user_filter')) {
-                $search_filter = "(&{" . Config::get('auth_ad_user_filter') . $search_filter . ")";
-            }
-            $search = ldap_search($this->ldap_connection, Config::get('auth_ad_base_dn'), $search_filter, array('samaccountname','displayname','objectsid','mail'));
-            $results = ldap_get_entries($this->ldap_connection, $search);
-
-            foreach ($results as $result) {
-                if (isset($result['samaccountname'][0])) {
-                    $userid = preg_replace(
-                        '/.*-(\d+)$/',
-                        '$1',
-                        $this->sidFromLdap($result['objectsid'][0])
-                    );
-
-                    // don't make duplicates, user may be member of more than one group
-                    $userhash[$result['samaccountname'][0]] = array(
-                        'realname' => $result['displayName'][0],
-                        'user_id'  => $userid,
-                        'email'    => $result['mail'][0]
-                    );
-                }
-            }
-        }
-
-        foreach (array_keys($userhash) as $key) {
-            $userlist[] = array(
-                'username' => $key,
-                'realname' => $userhash[$key]['realname'],
-                'user_id'  => $userhash[$key]['user_id'],
-                'email'    => $userhash[$key]['email']
-            );
-        }
-
-        return $userlist;
-    }
-
-    protected function getFullname($username)
-    {
-        $attributes = array('name');
-        $result = ldap_search(
-            $this->ldap_connection,
-            Config::get('auth_ad_base_dn'),
-            ActiveDirectoryAuthorizer::userFilter($username),
-            $attributes
-        );
-        $entries = ldap_get_entries($this->ldap_connection, $result);
-        if ($entries['count'] > 0) {
-            $membername = $entries[0]['name'][0];
-        } else {
-            $membername = $username;
-        }
-
-        return $membername;
-    }
-
-
-    public function getGroupList()
-    {
-        $ldap_groups   = array();
-
-        // show all Active Directory Users by default
-        $default_group = 'Users';
-
-        if (Config::has('auth_ad_group')) {
-            if (Config::get('auth_ad_group') !== $default_group) {
-                $ldap_groups[] = Config::get('auth_ad_group');
-            }
-        }
-
-        if (!Config::has('auth_ad_groups') && !Config::has('auth_ad_group')) {
-            $ldap_groups[] = $this->getDn($default_group);
-        }
-
-        foreach (Config::get('auth_ad_groups') as $key => $value) {
-            $ldap_groups[] = $this->getDn($key);
-        }
-
-        return $ldap_groups;
-    }
-
-    protected function getDn($samaccountname)
-    {
-        $attributes = array('dn');
-        $result = ldap_search(
-            $this->ldap_connection,
-            Config::get('auth_ad_base_dn'),
-            ActiveDirectoryAuthorizer::groupFilter($samaccountname),
-            $attributes
-        );
-        $entries = ldap_get_entries($this->ldap_connection, $result);
-        if ($entries['count'] > 0) {
-            return $entries[0]['dn'];
-        } else {
-            return '';
-        }
-    }
-
-    protected function getCn($dn)
-    {
-        preg_match('/[^,]*/', $dn, $matches, PREG_OFFSET_CAPTURE, 3);
-        return $matches[0][0];
-    }
-
-    protected function sidFromLdap($sid)
-    {
-        $sidHex = unpack('H*hex', $sid);
-        $subAuths = unpack('H2/H2/n/N/V*', $sid);
-        $revLevel = hexdec(substr($sidHex, 0, 2));
-        $authIdent = hexdec(substr($sidHex, 4, 12));
-        return 'S-'.$revLevel.'-'.$authIdent.'-'.implode('-', $subAuths);
+        return $this->ldap_connection;
     }
 }
