+++ LibreNMS/Authentication/ActiveDirectoryAuthorizer.php
@@ -10,6 +10,8 @@ use LibreNMS\Exceptions\AuthenticationException;
 
 class ActiveDirectoryAuthorizer extends AuthorizerBase
 {
+    use ADUtils;
+
     protected static $CAN_UPDATE_PASSWORDS = 0;
 
     protected $ldap_connection;
@@ -58,8 +60,6 @@ class ActiveDirectoryAuthorizer extends AuthorizerBase
     protected function userInGroup($username, $groupname)
     {
         // check if user is member of the given group or nested groups
-
-
         $search_filter = "(&(objectClass=group)(cn=$groupname))";
 
         // get DN for auth_ad_group
@@ -93,7 +93,7 @@ class ActiveDirectoryAuthorizer extends AuthorizerBase
             Config::get('auth_ad_base_dn'),
             // add 'LDAP_MATCHING_RULE_IN_CHAIN to the user filter to search for $username in nested $group_dn
             // limiting to "DN" for shorter array
-            "(&" . static::userFilter($username) . "(memberOf:1.2.840.113556.1.4.1941:=$group_dn))",
+            "(&" . $this->userFilter($username) . "(memberOf:1.2.840.113556.1.4.1941:=$group_dn))",
             array("DN")
         );
         $entries = ldap_get_entries($this->ldap_connection, $search);
@@ -108,7 +108,7 @@ class ActiveDirectoryAuthorizer extends AuthorizerBase
         $search = ldap_search(
             $this->ldap_connection,
             Config::get('auth_ad_base_dn'),
-            static::userFilter($username),
+            $this->userFilter($username),
             array('samaccountname')
         );
         $entries = ldap_get_entries($this->ldap_connection, $search);
@@ -155,7 +155,7 @@ class ActiveDirectoryAuthorizer extends AuthorizerBase
         $search = ldap_search(
             $this->ldap_connection,
             Config::get('auth_ad_base_dn'),
-            static::userFilter($username),
+            $this->userFilter($username),
             $attributes
         );
         $entries = ldap_get_entries($this->ldap_connection, $search);
@@ -200,164 +200,6 @@ class ActiveDirectoryAuthorizer extends AuthorizerBase
         }
 
         return array();
-    }
-
-
-    public function getUserlist()
-    {
-        $this->bind(); // make sure we called bind
-
-        $userlist = array();
-        $ldap_groups = $this->getGroupList();
-
-        foreach ($ldap_groups as $ldap_group) {
-            $search_filter = "(&(memberOf:1.2.840.113556.1.4.1941:=$ldap_group)(!(userAccountControl:1.2.840.113556.1.4.803:=2)))";
-            if (Config::get('auth_ad_user_filter')) {
-                $search_filter = "(&" . Config::get('auth_ad_user_filter') . $search_filter .")";
-            }
-            $attributes = array('samaccountname', 'displayname', 'objectsid', 'mail');
-            $search = ldap_search($this->ldap_connection, Config::get('auth_ad_base_dn'), $search_filter, $attributes);
-            $results = ldap_get_entries($this->ldap_connection, $search);
-
-            foreach ($results as $result) {
-                if (isset($result['samaccountname'][0])) {
-                    $userlist[$result['samaccountname'][0]] = $this->userFromAd($result);
-                }
-            }
-        }
-
-        return array_values($userlist);
-    }
-
-    /**
-     * Generate a user array from an AD LDAP entry
-     * Must have the attributes: objectsid, samaccountname, displayname, mail
-     * @internal
-     *
-     * @param $entry
-     * @return array
-     */
-    protected function userFromAd($entry)
-    {
-        return array(
-            'user_id' => $this->getUseridFromSid($this->sidFromLdap($entry['objectsid'][0])),
-            'username' => $entry['samaccountname'][0],
-            'realname' => $entry['displayname'][0],
-            'email' => isset($entry['mail'][0]) ? $entry['mail'][0] : null,
-            'descr' => '',
-            'level' => $this->getUserlevel($entry['samaccountname'][0]),
-            'can_modify_passwd' => 0,
-        );
-    }
-
-    protected function getEmail($username)
-    {
-        $this->bind(); // make sure we called bind
-
-        $attributes = array('mail');
-        $search = ldap_search(
-            $this->ldap_connection,
-            Config::get('auth_ad_base_dn'),
-            static::userFilter($username),
-            $attributes
-        );
-        $result = ldap_get_entries($this->ldap_connection, $search);
-        unset($result[0]['mail']['count']);
-        return current($result[0]['mail']);
-    }
-
-    protected function getFullname($username)
-    {
-        $this->bind(); // make sure we called bind
-
-        $attributes = array('name');
-        $result = ldap_search(
-            $this->ldap_connection,
-            Config::get('auth_ad_base_dn'),
-            static::userFilter($username),
-            $attributes
-        );
-        $entries = ldap_get_entries($this->ldap_connection, $result);
-        if ($entries['count'] > 0) {
-            $membername = $entries[0]['name'][0];
-        } else {
-            $membername = $username;
-        }
-
-        return $membername;
-    }
-
-
-    public function getGroupList()
-    {
-        $ldap_groups   = array();
-
-        // show all Active Directory Users by default
-        $default_group = 'Users';
-
-        if (Config::has('auth_ad_group')) {
-            if (Config::get('auth_ad_group') !== $default_group) {
-                $ldap_groups[] = Config::get('auth_ad_group');
-            }
-        }
-
-        if (!Config::has('auth_ad_groups') && !Config::has('auth_ad_group')) {
-            $ldap_groups[] = $this->getDn($default_group);
-        }
-
-        foreach (Config::get('auth_ad_groups') as $key => $value) {
-            $ldap_groups[] = $this->getDn($key);
-        }
-
-        return $ldap_groups;
-    }
-
-    protected function getDn($samaccountname)
-    {
-        $this->bind(); // make sure we called bind
-
-        $attributes = array('dn');
-        $result = ldap_search(
-            $this->ldap_connection,
-            Config::get('auth_ad_base_dn'),
-            static::groupFilter($samaccountname),
-            $attributes
-        );
-        $entries = ldap_get_entries($this->ldap_connection, $result);
-        if ($entries['count'] > 0) {
-            return $entries[0]['dn'];
-        } else {
-            return '';
-        }
-    }
-
-    protected function getCn($dn)
-    {
-        $dn = str_replace('\\,', '~C0mmA~', $dn);
-        preg_match('/[^,]*/', $dn, $matches, PREG_OFFSET_CAPTURE, 3);
-        return str_replace('~C0mmA~', ',', $matches[0][0]);
-    }
-
-    protected function getUseridFromSid($sid)
-    {
-        return preg_replace('/.*-(\d+)$/', '$1', $sid);
-    }
-
-    protected function sidFromLdap($sid)
-    {
-        $sidUnpacked = unpack('H*hex', $sid);
-        $sidHex = array_shift($sidUnpacked);
-        $subAuths = unpack('H2/H2/n/N/V*', $sid);
-        if (PHP_INT_SIZE <= 4) {
-            for ($i = 1; $i <= count($subAuths); $i++) {
-                if ($subAuths[$i] < 0) {
-                    $subAuths[$i] = $subAuths[$i] + 0x100000000;
-                }
-            }
-        }
-        $revLevel = hexdec(substr($sidHex, 0, 2));
-        $authIdent = hexdec(substr($sidHex, 4, 12));
-        return 'S-'.$revLevel.'-'.$authIdent.'-'.implode('-', $subAuths);
     }
 
     /**
@@ -434,29 +276,9 @@ class ActiveDirectoryAuthorizer extends AuthorizerBase
         ldap_set_option($this->ldap_connection, LDAP_OPT_PROTOCOL_VERSION, 3);
     }
 
-    public static function userFilter($username)
+    protected function getConnection()
     {
-        // don't return disabled users
-        $user_filter = "(&(samaccountname=$username)(!(useraccountcontrol:1.2.840.113556.1.4.803:=2))";
-
-        $extra = Config::get('auth_ad_user_filter');
-        if ($extra) {
-            $user_filter .= $extra;
-        }
-        $user_filter .= ')';
-
-        return $user_filter;
-    }
-
-    public static function groupFilter($groupname)
-    {
-        $group_filter = "(samaccountname=$groupname)";
-
-        $extra = Config::get('auth_ad_group_filter');
-        if ($extra) {
-            $group_filter = "(&$extra$group_filter)";
-        }
-
-        return $group_filter;
+        $this->bind(); // make sure connected and bound
+        return $this->ldap_connection;
     }
 }
